+++ main.py
@@ -1,6 +1,7 @@
-from qbee import *
 from examples import *
 from functools import partial
+
+from qbee.experimental import polynomialize_and_quadratize_pde
 
 if __name__ == '__main__':
     x, c = functions("x, c")
@@ -12,7 +13,7 @@ if __name__ == '__main__':
     ]
 
     no_quad_pruning = partial(pruning_by_nodes_without_quadratization_found, nodes_processed=1000)
-    res = polynomialize_and_quadratize(system, pruning_functions=[no_quad_pruning, *default_pruning_rules])
+    res = polynomialize_and_quadratize_pde(system, pruning_functions=[no_quad_pruning, *default_pruning_rules])
     assert res is not None
     print(res)
